Move RecordingCluster out of public API
