+++ docker-compose-rule-junit4/src/main/java/com/palantir/docker/compose/RecordingCluster.java
@@ -14,18 +14,21 @@
  * limitations under the License.
  */
 
-package com.palantir.docker.compose.connection;
+package com.palantir.docker.compose;
 
 import com.google.common.collect.Sets;
+import com.palantir.docker.compose.connection.Cluster;
+import com.palantir.docker.compose.connection.Container;
+import com.palantir.docker.compose.connection.ContainerCache;
 import java.io.IOException;
 import java.util.List;
 import java.util.Set;
 
-public class RecordingCluster extends Cluster {
+final class RecordingCluster extends Cluster {
     private final Cluster realCluster;
     private final Set<String> recordedContainerNames = Sets.newConcurrentHashSet();
 
-    public RecordingCluster(Cluster realCluster) {
+    RecordingCluster(Cluster realCluster) {
         this.realCluster = realCluster;
     }
